+++ client/cmd/rsrc_windows_386.syso
