+++ client/cmd/rsrc_windows_amd64.syso
